requirements.txt
python-telegram-bot>=21.0
anthropic>=0.40.0
gspread>=6.0
google-auth>=2.30
apscheduler>=3.10
python-dotenv>=1.0
